requirements.txt
pandas
langchain
langchain-core
langchain-community
langchain-experimental
langchain-openai
langchainhub
langgraph
langsmith
playwright
python-dotenv
bs4